refactor(dashboard): move run-sql logic to a custom hook
