+++ dashboard/src/features/database/dataGrid/hooks/useRunSQL/index.ts
@@ -1,0 +1,2 @@
+export * from './useRunSQL';
+export { default as useRunSQL } from './useRunSQL';
